refactor(portaltransform): 优化代码结构并添加中文翻译
- 重构了 EmiClientPlugin 和 EmiRecipe 类的导入路径- 更新了中文翻译文件，添加了多个 tooltip 相关的翻译
